Issue 30 fixed: change default password database behavior.
If a file with name that ends with ".rdf" exists in
the directory where pwmje was launched, pwmje will attempt to launch it.
If the file is invalid, pwmje will silently fail (since it is
entirely possible to have other rdf files in the same directory). 
Additionally if no file is found, it will attempt to load the
following files:
      o $HOME/passwordmaker.rdf
      o $HOME/pwmje.rdf
      o $HOME/.pwmje.rdf
      o $HOME/.passwordmakerrc

Again, any failure will be silent.

The Usage section of the wiki has also been updated with this
information.